+++ package.json
@@ -31,6 +31,9 @@
     "react-native-safe-area-context": "4.6.3",
     "react-native-screens": "~3.22.0",
     "react-native-web": "~0.19.6",
+    "react-redux": "^8.1.2",
+    "redux": "^4.2.1",
+    "redux-saga": "^1.2.3",
     "styled-components": "^6.0.7"
   },
   "devDependencies": {
